refactor: core:ui 패키지 정리
