--- core/ui/src/main/java/com/tjyg/core/ui/MviContract.kt
@@ -1,7 +1,0 @@
-package com.tjyg.core.ui
-
-interface UiState
-
-interface UiIntent
-
-interface UiSideEffect
