+++ core/ui/src/main/java/com/tjyg/core/ui/base/BaseViewModel.kt
@@ -1,7 +1,10 @@
-package com.tjyg.core.ui
+package com.tjyg.core.ui.base
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.tjyg.core.ui.mvi.UiIntent
+import com.tjyg.core.ui.mvi.UiSideEffect
+import com.tjyg.core.ui.mvi.UiState
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asSharedFlow
