+++ core/ui/src/main/java/com/tjyg/core/ui/mvi/UiIntent.kt
@@ -1,0 +1,3 @@
+package com.tjyg.core.ui.mvi
+
+interface UiIntent
