+++ core/ui/src/main/java/com/tjyg/core/ui/mvi/UiSideEffect.kt
@@ -1,0 +1,3 @@
+package com.tjyg.core.ui.mvi
+
+interface UiSideEffect
