+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserIntent.kt
@@ -1,6 +1,6 @@
 package com.teamyg.feature.sample
 
-import com.tjyg.core.ui.UiIntent
+import com.tjyg.core.ui.mvi.UiIntent
 
 sealed interface UserIntent : UiIntent {
     object LoadUser : UserIntent
