+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserSideEffect.kt
@@ -1,6 +1,6 @@
 package com.teamyg.feature.sample
 
-import com.tjyg.core.ui.UiSideEffect
+import com.tjyg.core.ui.mvi.UiSideEffect
 
 sealed interface UserSideEffect : UiSideEffect {
     data class ShowToast(val message: String) : UserSideEffect
