+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserState.kt
@@ -1,5 +1,5 @@
 package com.teamyg.feature.sample
 
-import com.tjyg.core.ui.UiState
+import com.tjyg.core.ui.mvi.UiState
 
 data class UserState(val isLoading: Boolean = false, val userName: String = "", val error: String? = null) : UiState
